Remove gdal package req from tests
It was only used to test the blocksize of the overviews of COGs being
created, since they had used 128 as a default. However, since gdal 3.1
the default has changed to be the same as the blocksize as the main
image. This should mean this is no longer and issue, and we no longer
need to install gdal just to run one test.